refactor: CQS 패턴 적용 및 컨버터 구현으로 코드 구조 개선
