+++ src/main/java/com/adit/backend/domain/place/dto/request/CommonPlaceRequestDto.java
@@ -6,10 +6,12 @@ import java.math.BigDecimal;
 import com.adit.backend.domain.place.entity.CommonPlace;
 
 import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 
 /**
  * DTO for {@link CommonPlace}
  */
+@AllArgsConstructor
 public record CommonPlaceRequestDto(@NotNull(message = "Place name must not be null") String placeName,
 									BigDecimal latitude, BigDecimal longitude, String addressName,
 									String roadAddressName, String subCategory, String url)
