+++ src/main/java/com/adit/backend/domain/place/entity/UserPlace.java
@@ -22,35 +22,35 @@ import lombok.Setter;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class UserPlace extends BaseEntity {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", nullable = false)
-    private User user;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id", nullable = false)
+	private User user;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "common_place_id", nullable = false)
-    private CommonPlace commonPlace;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "common_place_id", nullable = false)
+	private CommonPlace commonPlace;
 
-    private String memo;
-    private Boolean visited;
+	private String memo;
+	private Boolean visited;
 
-    @Builder
-    public UserPlace(User user, CommonPlace commonPlace,String memo, Boolean visited){
-        this.user = user;
-        this.commonPlace = commonPlace;
-        this.memo = memo;
-        this.visited = visited;
-    }
+	@Builder
+	public UserPlace(User user, CommonPlace commonPlace, String memo, Boolean visited) {
+		this.user = user;
+		this.commonPlace = commonPlace;
+		this.memo = memo;
+		this.visited = visited;
+	}
 
-    public void updatedMemo(String memo){
-        this.memo = memo;
-    }
+	public void updatedMemo(String memo) {
+		this.memo = memo;
+	}
 
-    public void updatedVisited(){
-        this.visited = true;
-    }
+	public void updatedVisited() {
+		this.visited = true;
+	}
 
 }
